requirements.txt
asgiref==3.4.1
certifi==2021.5.30
coverage==5.5
crispy-bootstrap5==0.4
dj-database-url==0.5.0
Django==3.2.5
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-heroku==0.3.1
django-model-utils==4.1.1
factory-boy==3.2.0
Faker==8.10.1
gunicorn==20.1.0
Pillow==8.3.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2021.1
sentry-sdk==1.3.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.6
whitenoise==5.3.0
